+++ src/pages/index.ts
@@ -1,3 +1,3 @@
 export { default as Login } from './Login/Login';
-export { default as SingIn } from './SingIn/SingIn';
+export { default as SingIn } from './SignIn/SignIn';
 export { default as Home } from './Home/Home';
